event-wall-api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
supabase==2.28.3
Pillow==12.2.0
google-cloud-storage==3.1.0
google-cloud-monitoring==2.30.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytest==8.3.5
pytest-cov==6.1.0
